WIP: multiple guest account locks:
- guest profiles are invisible
- guests are not part of the forum
- user dashboard is locked and redirects to the linked guest group
- guests never have write access to anything
