+++ cosinnus/templates/cosinnus/group/group_detail.html
@@ -679,7 +679,9 @@
                                                         {% endif %}
                                                     </span>
                                                     <a class="username-name" href="{% url 'cosinnus:profile-detail' username=member.username %}">
-                                                        {% if not member_active %}
+                                                        {% if member.is_guest %}
+                                                            <i>({% trans "Guest" %})</i>
+                                                        {% elif not member_active %}
                                                             <i>({% trans "User account inactive" %})</i>
                                                         {% endif %}
                                                         {{ member|full_name }}
